Add files via upload
A classification problem is when the output variable is a category, such as “red” or “blue” or “disease” and “no disease”. A classification model attempts to draw some conclusion from observed values. Given one or more inputs a classification model will try to predict the value of one or more outcomes.

Here is an example of classification in which we are performing classification on the iris dataset using RandomForestClassifier in python.